+++ codec/legacy/codec.go
@@ -22,11 +22,11 @@ func init() {
 // PrivKeyFromBytes unmarshals private key bytes and returns a PrivKey
 func PrivKeyFromBytes(privKeyBytes []byte) (privKey cryptotypes.PrivKey, err error) {
 	err = Cdc.Unmarshal(privKeyBytes, &privKey)
-	return
+	return privKey, err
 }
 
 // PubKeyFromBytes unmarshals public key bytes and returns a PubKey
 func PubKeyFromBytes(pubKeyBytes []byte) (pubKey cryptotypes.PubKey, err error) {
 	err = Cdc.Unmarshal(pubKeyBytes, &pubKey)
-	return
+	return pubKey, err
 }