+++ crypto/keys/secp256k1/secp256k1_cgo_test.go
@@ -1,5 +1,4 @@
 //go:build libsecp256k1_sdk
-// +build libsecp256k1_sdk
 
 package secp256k1
 
@@ -7,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/magiconair/properties/assert"
-
 	"github.com/stretchr/testify/require"
 )