+++ x/gov/client/utils/query.go
@@ -161,7 +161,8 @@ func QueryVoteByTxQuery(clientCtx client.Context, params v1.QueryVoteParams) ([]
 }
 
 // QueryProposerByTxQuery will query for a proposer of a governance proposal by ID.
-// Deprecated: Should not be used, as not always accurate. It will be removed in v0.51.
+//
+// Deprecated: Should not be used, as not always accurate. It will be removed in v0.54.
 func QueryProposerByTxQuery(clientCtx client.Context, proposalID uint64) (Proposer, error) {
 	q := fmt.Sprintf("%s.%s='%d'", types.EventTypeSubmitProposal, types.AttributeKeyProposalID, proposalID)
 	searchResult, err := authtx.QueryTxsByEvents(clientCtx, defaultPage, defaultLimit, q, "")